Created Photograph handler script
The script currently choses either the false or correct image list, shuffles the list, then chooses the first item from that list. It should check if the image has already been displayed and put it into a list in which it wont display again. It also has a function to return the current image to display.